requirements.txt
ultralytics
nxt-python
opencv-python
numpy
pyusb
